Fixed some file mixups
