+++ frontend/src/redux/slices/profileSlice.js
@@ -1,4 +1,4 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, current } from '@reduxjs/toolkit';
 
 const profileSlice = createSlice({
     name: 'profileSlice',
@@ -8,10 +8,14 @@ const profileSlice = createSlice({
     reducers: {
         createProfile: (state, action) => {
             state.profiles.push(action.payload);
+            console.log(current(state.profiles));
         },
         editProfile: (state, action) => {
             const index = state.profiles.findIndex((el) => el.username === action.payload.username);
+            console.log('In edit');
+            console.log(index);
             state.profiles[index] = action.payload;
+            console.log(current(state.profiles));
         },
     },
 });
